Use usize for element_width
Allows easy use of std::mem::size_of for picking element sizes instead
of hardcoding them

+++ src/compute/shaders.rs
@@ -422,7 +422,7 @@ pub enum IntermediateDataDescription {
         /// Dimensions of the buffer
         dim: BufferDim,
         /// Element width in bytes
-        element_width: u8,
+        element_width: usize,
     },
 }
